+++ scripts/ac5e-constants.mjs
@@ -2,6 +2,7 @@ export default class Constants {
 	static MODULE_ID = 'automated-conditions-5e';
 	static MODULE_NAME = 'Automated Conditions 5e';
 	static MODULE_NAME_SHORT = 'ac5e';
+	static ACTIVE_EFFECT_CHANGE_TYPE = 'ac5e';
 	static GM_EFFECT_DELETIONS = 'automated-conditions-5e.gmEffectDeletions';
 	static GM_DOCUMENT_UPDATES = 'automated-conditions-5e.gmDocumentUpdates';
 	static GM_STATUS_TOGGLES = 'automated-conditions-5e.gmStatusToggles';